==> code/3. Text preprocessing (Local).ipynb <==
## Package load

import os
import numpy as np
import pandas as pd

## Setting path

train = pd.read_csv("../../KEMDy20_v1_1/Splitting/Train.csv")
test = pd.read_csv("../../KEMDy20_v1_1/Splitting/Test.csv")
train_origin = pd.read_csv("../../KEMDy20_v1_1/Splitting/Train_Origin.csv")
test_origin = pd.read_csv("../../KEMDy20_v1_1/Splitting/Test_Origin.csv")
aug_an_file_path = "../../AI_hub/annotation/5차년도_2차.csv"

## Path setting (AI Hub)

aug_an_file = pd.read_csv(aug_an_file_path, encoding="cp949")

## Merge text to train(KEMDy20 + AI Hub)

train_text_list = []
for path in train["Path"]:
    if path.split("/")[2]=="AI_hub":
        wav_id = path.split("/")[4][:-4]
        train_text_list.append(aug_an_file[aug_an_file["wav_id"]==wav_id]["발화문"].values[0])
    else:
        with open(path[:-4]+".txt", "r") as f:
            text = f.read()
            train_text_list.append(text)
            
train["Text"] = train_text_list

## Merge text to train(KEMDy20)

train_origin_text_list = []
for path in train_origin["Path"]:
    with open(path[:-4]+".txt", "r") as f:
        text = f.read()
        train_origin_text_list.append(text)
        
train_origin["Text"] = train_origin_text_list

## Metge text to test(KEMDy20 + AI Hub)

test_text_list = []
for path in test["Path"]:
    if path.split("/")[2]=="AI_hub":
        wav_id = path.split("/")[4][:-4]
        test_text_list.append(aug_an_file[aug_an_file["wav_id"]==wav_id]["발화문"].values[0])
    else:
        with open(path[:-4]+".txt", "r") as f:
            text = f.read()
            test_text_list.append(text)

test["Text"] = test_text_list

## Metge text to test(KEMDy20)

test_origin_text_list = []
for path in test_origin["Path"]:
    with open(path[:-4]+".txt", "r") as f:
        text = f.read()
        test_origin_text_list.append(text)
        
test_origin["Text"] = test_origin_text_list

## Save path setting

save_file_path = "../../KEMDy20_v1_1/Text/"
if not os.path.exists(save_file_path):
    os.mkdir(save_file_path)

## Save Train, Test text file (KEMDy20 + AI Hub)

train[["Text", "Emotion"]].to_csv(save_file_path+"Train.csv", index=False, encoding="euc-kr")
test[["Text", "Emotion"]].to_csv(save_file_path+"Test.csv", index=False, encoding="euc-kr")

## Save Train, Test text file (KEMDy20)

train_origin[["Text", "Emotion"]].to_csv(save_file_path+"Train_Origin.csv", index=False, encoding="euc-kr")
test_origin[["Text", "Emotion"]].to_csv(save_file_path+"Test_Origin.csv", index=False, encoding="euc-kr")
