==> code/8. Ensemble prediction (Local).ipynb <==
## Package load

import random
import numpy as np
import pandas as pd

from keras.utils import to_categorical
from sklearn.preprocessing import LabelEncoder
from sklearn.metrics import classification_report

## Load pred probability

wav2vec_pred = np.load("../../KEMDy20_v1_1/pred/wav2vec.npz")["predict_prob"]
wav2vec_R_pred = np.load("../../KEMDy20_v1_1/pred/wav2vec_R.npz")["predict_prob"]
kobert_03_pred = np.load("../../KEMDy20_v1_1/pred/KoBERT_03.npz")["predict_prob"]
kobert_05_pred = np.load("../../KEMDy20_v1_1/pred/KoBERT_05.npz")["predict_prob"]
kobert_07_pred = np.load("../../KEMDy20_v1_1/pred/KoBERT_07.npz")["predict_prob"]
test_y = pd.read_csv("../../KEMDy20_v1_1/Splitting/Test.csv").Emotion.values

## Label encoding

le = LabelEncoder()
le.fit(test_y)
test_y = le.transform(test_y)

print(wav2vec_pred.shape)
print(wav2vec_R_pred.shape)
print(kobert_03_pred.shape)
print(kobert_05_pred.shape)
print(kobert_07_pred.shape)
print(test_y.shape)

## Ensemble

wav2vec_bert_test_pred = (wav2vec_pred*0.25 + wav2vec_R_pred*0.25 + kobert_03_pred*0.16666666 + kobert_05_pred*0.16666666 + kobert_07_pred*0.16666666)

## Probability to argmax

wav2vec_y = np.argmax(wav2vec_pred, axis=1)
wav2vec_R_y = np.argmax(wav2vec_R_pred, axis=1)
kobert_03_y = np.argmax(kobert_03_pred, axis=1)
kobert_05_y = np.argmax(kobert_05_pred, axis=1)
kobert_07_y = np.argmax(kobert_07_pred, axis=1)
wav2vec_bert_ensemble = np.argmax(wav2vec_bert_test_pred, axis=1)

## wav2vec2 emotion recognition (forward)

print(classification_report(test_y, wav2vec_y, target_names=le.classes_, digits=4))

## wav2vec2 emotion recognition (backward)

print(classification_report(test_y, wav2vec_R_y, target_names=le.classes_, digits=4))

## KoBERT emotion recognition (dr_rate=0.3)

print(classification_report(test_y, kobert_03_y, target_names=le.classes_, digits=4))

## KoBERT emotion recognition (dr_rate=0.5)

print(classification_report(test_y, kobert_05_y, target_names=le.classes_, digits=4))

## KoBERT emotion recognition (dr_rate=0.7)

print(classification_report(test_y, kobert_07_y, target_names=le.classes_, digits=4))

## Ensemble emotion recognition

print(classification_report(test_y, wav2vec_bert_ensemble, target_names=le.classes_, digits=4))
